requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pytesseract
pillow
python-multipart
python-dotenv
pydantic